pyproject.toml
[tool.poetry]
name = "build-dagger"
version = "0.1.0"
description = ""
authors = ["pythoninthegrass <4097471+pythoninthegrass@users.noreply.github.com>"]
license = "Unlicense"
readme = "README.md"
packages = [{ include = "main", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
anyio = "^4.6.2.post1"
dagger-io = "^0.13.7"
python-decouple = "^3.8"
names = "^0.3.0"

[tool.dagger]
base-image = "python:3.11-slim"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
